Update IndexerMain for JSONL file (BROKEN)
This patch works with the tree at commit 4b582f5 but not with the
current HEAD.

The input at the moment has to be one JSONL file on the local
filesystem.

When linked with the dependencies and other classes specified in the
current HEAD, you get this exception:

```
[error] (run-main-0) org.apache.spark.SparkException: Job aborted due to
stage failure: [...] java.lang.NoClassDefFoundError:
org/apache/commons/httpclient/HttpConnectionManager
```
... Doesn't matter whether you update build.sbt with spark-core 2.0 or
2.1, or the latest version of the elasticsearch-spark library; same
error.

The changes to IndexerMain.scala work on top of the tree at commmit 4b582f5
with Elasticsearch 5 (keeping the Elasticsearch version consistent for the
time being until the issue above can be figured out).

TODO:
* Fix elasticsearch-spark dependency?
* Make work with Elasticsearch 6
* Have it work with s3:// URIs
* Have it work with an S3 "folder" or filesystem directory